/search/ undocumented ?tag= and ?exclude.tag= params
e.g. https://simonwillison.com/search/?q=django&tag=quora

- Search for Django, return results with tag quora

Or...

https://simonwillison.com/search/?q=jquery&exclude.tag=jquery

- Find jquery results that do NOT have tag jquery

You can stack these up like so:

https://simonwillison.com/search/?q=jquery&tag=javascript&tag=jquery

BUT this does not yet do what I'd like it to do - this will OR the tags,
showing everything matching result that is tagged javascript or jquery. I plan
for this to AND things in the future.

I also want to add a faceted UI that exposes these options.