Mechanism for logging hits from Atom feeds in Google Analytics
I want some kind of indication of if anyone is using my Atom feeds.

To support this, I'm going to rewrite the links in my feeds to
look something like this:

https://simonwillison.net/2009/Nov/23/node/#atom-everything

Where #atom-everything (or #atom-entries or #atom-blogmarks) tells me which
feed the link was in.

I've added some JavaScript which checks for these fragments on page load and,
if one is found, sends it as a custom dimension to Google Analytics. It then
uses HTML5 history.replaceState to eliminate the parameter from the URL,
because they're ugly and I don't want them spreading into other links to the
page.