Fixed some encoding errors in my XHTML-handling tags
As a result, first paragraph is now more reliably displayed. Previously it was
frequently causing an exception which Django's template system then supressed.

The pattern here is that any time you feed data to ElementTree.fromstring you
need to first encode it as UTF-8. When you are done manipulating the XML you
can convert it back to a unicode string with the .decode() method.