Mechanism for counting feed subscribers based on User-Agent
A number of feed reader services expose subscriber counts in their user-agent
string, like this:

* NewsGatorOnline/2.0 (http://www.newsgator.com; 47 subscribers)
* NewsBlur Feed Fetcher - 496 subscribers - ...
* Feedbin feed-id:379793 - 21 subscribers

The new feedstats app looks out for this pattern of user-agent string and
stores a row in the new SubscriberCount model every time it spots a new number
of subscribers. This will allow me to track subscriber growth over time.