Massive clean-up of |safe, |escape and XHTML processing code
I think I originally wrote these templates back before Django had auto-
escaping! As a result, they were cluttered with calls to |safe and |escape and
the logic was getting pretty confusing.

I've managed to eliminate every single call to |escape and all but two
(necessary) calls to |safe.

Almost all of the content on my blog should be treated as regular strings
without markup - the the exception of the entry bodies themselves, which are
stored (and validated) as XHTML fragments.

Previously, my various template filters like |typography and |first_paragraph
would each take a string of (hopefully) XHTML, parse that into ElementTree,
process it in some way and then serialize it back to a string again. This
could happen dozens of times on a single page.

I've introduced a new XhtmlString class which allows me to execute that
parsing stage just once, and then pass an object along to each of the filters
that they can then work with. The __str__ method knows how to serialize back
to XML.

I've also introduced a |xhtml filter which expects a string of XHTML and turns
it into one of those objects, treating the markup as markup. Without this
filter, a string will be assumed to be e.g. a title with ampersands in it so
the string will be escaped before being handed to ElementTree.